+++ hooks/post_gen_project.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+
+import os
+
+if __name__ == "__main__":
+    print("Executing 'uv sync'")
+    os.system("uv sync")
+    print("Your Python package project has been created successfully!")